seperate test fixture objects into another file
